requirements.txt
numpy
scipy
scikit-tensor-py3
absl-py
flopco-pytorch
